+++ ricaun.Revit.UI.Tests/Panels/RevitTabPanelTests.cs
@@ -8,48 +8,50 @@ namespace ricaun.Revit.UI.Tests.Panels
         private const string TabName = "RevitTabPanelTests";
         private const string PanelName = "Example";
         private static RibbonPanel ribbonPanel;
+        private UIControlledApplication application;
 
         [SetUp]
         public void SetUp(UIControlledApplication application)
         {
+            this.application = application;
             ribbonPanel = application.CreatePanel(TabName, PanelName);
             ribbonPanel.Remove();
             System.Console.WriteLine($"RibbonTab: {ribbonPanel.GetRibbonTab().Id}");
         }
 
-        private bool ContainPanel(UIControlledApplication application, RibbonPanel panel)
+        private bool ContainPanel(RibbonPanel panel)
         {
             var tabName = panel.GetRibbonTab().Id;
             return application.GetRibbonPanels(tabName).Contains(panel);
         }
 
         [Test]
-        public void CreatePanel(UIControlledApplication application)
+        public void CreatePanel()
         {
             var panel = application.CreatePanel(TabName, PanelName);
             foreach (var item in RibbonTabExtension.GetRibbonTabsDictionary(panel.GetRibbonTab()))
             {
                 System.Console.WriteLine(item);
             }
-            Assert.IsTrue(ContainPanel(application, panel), "Contain Panel");
+            Assert.IsTrue(ContainPanel(panel), "Contain Panel");
             panel?.Remove();
-            Assert.IsFalse(ContainPanel(application, panel), "Not Contain Panel");
+            Assert.IsFalse(ContainPanel(panel), "Not Contain Panel");
         }
 
         [Test]
-        public void CreatePanel_10Times(UIControlledApplication application)
+        public void CreatePanel_10Times()
         {
             for (int i = 0; i < 10; i++)
             {
                 var panel = application.CreatePanel(TabName, PanelName);
-                Assert.IsTrue(ContainPanel(application, panel), "Contain Panel");
+                Assert.IsTrue(ContainPanel(panel), "Contain Panel");
                 panel?.Remove();
-                Assert.IsFalse(ContainPanel(application, panel), "Not Contain Panel");
+                Assert.IsFalse(ContainPanel(panel), "Not Contain Panel");
             }
         }
 
         [TearDown]
-        public void TearDown(UIControlledApplication application)
+        public void TearDown()
         {
             if (!string.IsNullOrEmpty(TabName))
                 ribbonPanel.GetRibbonTab().Remove();